feat(backend): Flyway e dinheiro em BigDecimal
Primeira etapa da liquidacao automatica via carteira. Ainda nao muda
comportamento: prepara o terreno tirando o dinheiro de ponto flutuante e
passando a propriedade do schema para migracoes versionadas.

Flyway
- flyway-core + flyway-database-postgresql (o suporte por banco saiu do
  core no Flyway 10).
- Producao: ddl-auto passa de "update" para "validate". O schema deixa de
  ser efeito colateral do boot e passa a ser o que as migracoes dizem.
- Dev continua H2/create-drop, com spring.flyway.enabled=false EXPLICITO:
  o flyway-core no classpath liga o Flyway sozinho e as migracoes sao SQL
  PostgreSQL, que o H2 nao entende.

Migracoes
- V1__baseline.sql: o schema como estava antes do Flyway, gerado a partir
  das entidades em d4984cb com o dialeto PostgreSQL (nao escrito a mao).
  Com baseline-on-migrate esta migracao NAO roda no banco do Railway, que
  ja existe; serve para ambiente novo e para versionar o schema.
- V2__p0_scrum_17_20.sql: colunas/tabelas do P0 (latitude, longitude,
  endereco, expirado_em, notificacoes, indices, unique de avaliacao).
  Chegavam via ddl-auto=update, que agora nao roda mais.
  IDEMPOTENTE de proposito: nao da para inspecionar o banco do Railway
  daqui e nao se sabe se o deploy do P0 ja aplicou essas colunas. Com
  IF NOT EXISTS o resultado final e o mesmo nos dois cenarios.
- V3__dinheiro_para_numeric.sql: FLOAT(53) -> NUMERIC(12,2).

BigDecimal
- Carteira.saldoAtual/ganhosMensais, Transacao.valor, Turno.valorEstimado.
  precision=12/scale=2 nas entidades para casar com o NUMERIC(12,2) da
  V3 — sem isso o Hibernate assume numeric(38,2) e o validate reprova.
- Comparacao por compareTo, nunca equals: equals considera a escala e
  "20.0" nao e igual a "20.00". Soma com add/subtract.
- setScale(2, HALF_UP) so na borda de saida (DTOs e respostas de mapa).
- Propagado por CarteiraService, TurnoService, DashboardController,
  RelatorioController, SugestaoController e DataInitializer. Nos helpers
  de seed a assinatura segue double, com conversao na borda, para nao
  reescrever dezenas de chamadas de dado ficticio.

Verificacao
- mvn clean compile: BUILD SUCCESS (44 arquivos).
- Boot em dev conferido: sobe, H2 cria as colunas como numeric(12,2) e
  nenhuma linha de Flyway aparece no log.
- mvn clean test segue vermelho (14 testes, 2 failures, 3 errors), mas
  ja estava vermelho em origin/main antes desta mudanca — conferido em
  worktree limpo. TurnoServiceTest nao mocka TurnoInscricaoRepository nem
  NotificacaoService, dependencias que o P0 adicionou ao TurnoService.

+++ backend/src/main/java/com/motoshift/config/DataInitializer.java
@@ -13,6 +13,7 @@ import com.motoshift.repository.UsuarioRepository;
 import org.springframework.boot.CommandLineRunner;
 import org.springframework.stereotype.Component;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.LocalTime;
@@ -326,8 +327,8 @@ public class DataInitializer implements CommandLineRunner {
     private void criarCarteira(Long motoboyId, double saldo, double ganhos, String pix) {
         Carteira c = new Carteira();
         c.setMotoboyId(motoboyId);
-        c.setSaldoAtual(saldo);
-        c.setGanhosMensais(ganhos);
+        c.setSaldoAtual(BigDecimal.valueOf(saldo));
+        c.setGanhosMensais(BigDecimal.valueOf(ganhos));
         c.setChavePix(pix);
         carteiraRepo.save(c);
     }
@@ -344,7 +345,7 @@ public class DataInitializer implements CommandLineRunner {
         t.setRegiao(regiao);
         t.setDataInicio(inicio);
         t.setDataFim(fim);
-        t.setValorEstimado(valor);
+        t.setValorEstimado(BigDecimal.valueOf(valor));
         t.setRaioEntregaKm(raio);
         // SCRUM-18: sem coordenada os turnos de exemplo nao aparecem no filtro
         // por raio e a tela parece quebrada em desenvolvimento.
@@ -428,7 +429,7 @@ public class DataInitializer implements CommandLineRunner {
         tx.setMotoboyId(motoboyId);
         tx.setTurnoId(turnoId);
         tx.setTipo(tipo);
-        tx.setValor(valor);
+        tx.setValor(BigDecimal.valueOf(valor));
         tx.setDescricao(descricao);
         tx.setStatus(status);
         transacaoRepo.save(tx);
